+++ tests/testthat/test-embed-images.R
@@ -34,16 +34,15 @@ test_that("embed_images() works for non-ASCII dir", {
   # integrated test
   Rmd_with_some_settings <-
 'title: "title1"
-confluence_settings:
-  space_key: "space1"'
+output:
+  conflr::confluence_document:
+    space_key: "space1"'
   confl_upload_mock <- mockery::mock(NULL)
   do_confl_create_post_from_Rmd(confl_upload_mock, Rmd_with_some_settings, body = md_text)
   result2 <- mockery::mock_args(confl_upload_mock)[[1]]
 
   expect_equal(result2$html_text, html_text)
-  expect_equal(
-    embed_images(result2$html_text, result2$imgs, result2$imgs_realpath),
-    expected
-  )
+  expect_equal(result2$imgs, "%C3%B6/plot1.png")
+  expect_equal(result2$imgs_realpath, "\u00f6/plot1.png")
 })